pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-logpipe"
version = "1.3.0.post4"
description = "Move data around between Python services using Kafka and/or AWS Kinesis and Django Rest Framework serializers."
authors = [ "thelabnyc <thelabdev@thelabnyc.com>",]
readme = "README.rst"
homepage = "https://gitlab.com/thelabnyc/django-logpipe"
repository = "https://gitlab.com/thelabnyc/django-logpipe"
license = "ISC"

[[tool.poetry.packages]]
include = "logpipe"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
Django = ">=4.2"
djangorestframework = ">=3.14"
lru-dict = ">=1.1.6"

[tool.poetry.group.dev.dependencies]
flake8 = "*"
moto = "^4.2.14"
boto3 = "*"
psycopg2-binary = "*"
tox = "*"
pytz = "*"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
boto3-stubs = {extras = ["essential", "kinesis"], version = "^1.34.40"}
botocore-stubs = "^1.34.40"
djangorestframework-stubs = "^3.14.5"
msgpack-types = "^0.2.0"

[tool.poetry.group.kafka.dependencies]
# Install Kafka from github until they pulish a Py3.12 version on PyPi
# kafka-python = "*"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git", rev = "master"}

[tool.poetry.group.kinesis.dependencies]
boto3 = "*"

[tool.poetry.group.msgpack.dependencies]
msgpack = "*"

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[[tool.mypy.overrides]]
module = "kafka.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "logpipe.tests.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "sandbox.settings"

[tool.isort]
profile = "black"
from_first = true
